Update Configuration.h
This edit is made to solve the  awkward method actually used to find the right probing square. This edit let the user make use of the X and Y value printed on screen on the GUI and find the lower leftmost point and the rightmost top point of the square in wich to probe. The process is simple. 

-Home Axis and rise the Z to let the space for the probe
-Extend the probe
-Bring the probe where you want the leftmost front point. Use m114 to print the actual position write down X and Y value. 
-Write X and Y value in the POS_A_X and POS_A_Y 
-Bring the probe where you want the rightmost back point. Use m114 to print the actual position write down X and Y value. 
-Write X and Y value in the POS_B_X and POS_B_Y 
- Still , every value has  to be consistent. You cant probe a point that your probe can't reach.  
End. ( Explaination video incoming )